update scripts to typescript

+++ scripts/dashboard/clear-logs.ts
@@ -1,6 +1,6 @@
 
-const path = require('path');
-const fs = require('fs');
+import fs from 'fs';
+import path from 'path';
 
 const dashboardPath = path.join(__dirname, '../../dashboard');
 
@@ -13,5 +13,5 @@ const logNames = [
 
 for (const logName of logNames) {
     const logFullPath = path.join(dashboardPath, logName);
-    fs.rmSync(logFullPath);
+    if (fs.existsSync(logFullPath)) fs.rmSync(logFullPath);
 }
